+++ Pandora/Paradigm/Structure.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Structure (module Exports) where
 
-import Pandora.Paradigm.Structure.Variation.Substructure as Exports
-import Pandora.Paradigm.Structure.Variation.Nonempty as Exports
+import Pandora.Paradigm.Structure.These.Substructure as Exports
+import Pandora.Paradigm.Structure.These.Nonempty as Exports
 import Pandora.Paradigm.Structure.Rose as Exports
 import Pandora.Paradigm.Structure.Binary as Exports
 import Pandora.Paradigm.Structure.Stack as Exports