+++ Pandora/Paradigm/Structure/Rose.hs
@@ -15,8 +15,8 @@ import Pandora.Paradigm.Primary.Transformer.Construction (Construction (Construc
 import Pandora.Paradigm.Controlflow.Joint.Schemes.TU (TU (TU))
 import Pandora.Paradigm.Inventory.Store (Store (Store))
 import Pandora.Paradigm.Structure.Stack (Stack)
-import Pandora.Paradigm.Structure.Variation.Nonempty (Nonempty)
-import Pandora.Paradigm.Structure.Variation.Substructure (Substructure (Output, sub))
+import Pandora.Paradigm.Structure.These.Nonempty (Nonempty)
+import Pandora.Paradigm.Structure.These.Substructure (Substructure (Output, sub))
 
 type Rose = TU Covariant Covariant Maybe (Construction Stack)