+++ Pandora/Paradigm/Structure/Variation/Nonempty.hs
@@ -1,4 +1,4 @@
-module Pandora.Paradigm.Structure.Variation.Nonempty (Nonempty) where
+module Pandora.Paradigm.Structure.These.Nonempty (Nonempty) where
 
 -- | Type synonymous for at least one element data structure
 type family Nonempty (s :: * -> *) = (r :: * -> *) | r -> s